Move pglb::connect_compute back to proxy::connect_compute
